pyproject.toml
[project]
name = "personalized-learning-coach"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.5",
    "djangorestframework>=3.16.1",
    "faiss-cpu>=1.7.4",
    "google-genai>=1.30.0",
    "groq>=0.4.0",
    "json5>=0.12.1",
    "langchain-google-genai>=1.0.0",
    "manim>=0.19.0",
    "mistralai>=1.9.6",
    "numpy>=1.24.0",
    "openai>=1.99.9",
    "pypdf2>=3.0.0",
    "python-dotenv==0.21.0",
    "setuptools>=80.9.0",
    "tiktoken>=0.5.0",
]
